Update change log for versions updated by version policy enforcement.

+++ apps/rush/src/actions/VersionAction.ts
@@ -2,13 +2,19 @@
 // See LICENSE in the project root for license information.
 
 import { EOL } from 'os';
+import * as fsx from 'fs-extra';
+import * as path from 'path';
+import * as semver from 'semver';
 
 import {
   CommandLineFlagParameter,
   CommandLineStringParameter
 } from '@microsoft/ts-command-line';
 import {
-  IPackageJson
+  IPackageJson,
+  IChangeInfo,
+  ChangeType,
+  RushConfigurationProject
 } from '@microsoft/rush-lib';
 
 import RushCommandLineParser from './RushCommandLineParser';
@@ -16,6 +22,7 @@ import GitPolicy from '../utilities/GitPolicy';
 import { BaseRushAction } from './BaseRushAction';
 import { VersionManager } from '../utilities/VersionManager';
 import { Git } from '../utilities/Git';
+import ChangelogGenerator from '../utilities/ChangelogGenerator';
 
 export default class VersionAction extends BaseRushAction {
   private _parser: RushCommandLineParser;
@@ -77,21 +84,77 @@ export default class VersionAction extends BaseRushAction {
       const tempBranch: string = 'version/ensure-' + new Date().getTime();
       const git: Git = new Git(this._targetBranch.value);
 
-      // Make changes in temp branch.
-      git.checkout(tempBranch, true);
       const updatedPackages: Map<string, IPackageJson> = this._versionManager.ensure(this._versionPolicy.value);
+      if (updatedPackages) {
+        console.log(`${updatedPackages.size} packages are getting updated.`);
+        this._updateFiles(updatedPackages);
+        // Make changes in temp branch.
+        git.checkout(tempBranch, true);
 
-      // Stage, commit, and push the changes to remote temp branch.
-      git.addChanges();
-      git.commit();
-      git.push(tempBranch);
+        // Stage, commit, and push the changes to remote temp branch.
+        git.addChanges();
+        git.commit();
+        git.push(tempBranch);
 
-      // Now merge to target branch.
-      git.checkout(this._targetBranch.value);
-      git.pull();
-      git.merge(tempBranch);
-      git.push(this._targetBranch.value);
-      git.deleteBranch(tempBranch);
+        // Now merge to target branch.
+        git.checkout(this._targetBranch.value);
+        git.pull();
+        git.merge(tempBranch);
+        git.push(this._targetBranch.value);
+        git.deleteBranch(tempBranch);
+      }
     }
+  }
+
+  private _updateFiles(updatedPackages: Map<string, IPackageJson>): void {
+    updatedPackages.forEach((newPackageJson, packageName) => {
+      const rushProject: RushConfigurationProject = this.rushConfiguration.getProjectByName(packageName);
+      // Update package.json
+      const packagePath: string = path.join(rushProject.projectFolder, 'package.json');
+      fsx.writeFileSync(packagePath, JSON.stringify(newPackageJson, undefined, 2), 'utf8');
+
+      if (newPackageJson.version !== rushProject.packageJson.version) {
+        // If package version changes, add an entry to changelog
+        const change: IChangeInfo = this._createChangeInfo(newPackageJson, rushProject);
+        ChangelogGenerator.updateIndividualChangelog(change,
+          rushProject.projectFolder,
+          true);
+      }
+      // TODO: if only package dependency changes, add change file.
+    });
+  }
+
+  private _createChangeInfo(newPackageJson: IPackageJson,
+    rushProject: RushConfigurationProject
+  ): IChangeInfo {
+    const changeType: ChangeType = this._getChangeType(rushProject.packageJson.version,
+      newPackageJson.version);
+    // TODO: need to absorb all existing change files
+    return {
+      changeType: changeType,
+      newVersion: newPackageJson.version,
+      packageName: newPackageJson.name,
+      changes: [
+        {
+          changeType: changeType,
+          comment: `Version bump to ${newPackageJson.version}`,
+          newVersion: newPackageJson.version,
+          packageName: newPackageJson.name
+        }
+      ]
+    };
+  }
+
+  private _getChangeType(oldVersionString: string, newVersionString: string): ChangeType {
+    const diff: string = semver.diff(oldVersionString, newVersionString);
+    let changeType: ChangeType = ChangeType.none;
+    if (diff === 'major') {
+      changeType = ChangeType.major;
+    } else if (diff === 'minor') {
+      changeType = ChangeType.minor;
+    } else if (diff === 'patch') {
+      changeType = ChangeType.patch;
+    }
+    return changeType;
   }
 }
